extract message component and utils
